Separate the stmt and expr more clearly
