+++ interpreter/src/ast/expr/assign.rs
@@ -1,6 +1,6 @@
 use crate::{scanner::token::Token, utils::error::Error};
 
-use super::expr::{self, Expr};
+use super::Expr;
 
 pub struct Assign<R> {
     pub name: Token,
@@ -22,7 +22,7 @@ impl<R> Assign<R> {
 }
 
 impl<R: 'static> Expr<R> for Assign<R> {
-    fn accept(&self, visitor: &mut dyn expr::Visitor<R>) -> Result<R, Error> {
+    fn accept(&self, visitor: &mut dyn super::Visitor<R>) -> Result<R, Error> {
         visitor.visit_assign_expr(self)
     }
