+++ interpreter/src/ast/expr/binary.rs
@@ -1,6 +1,6 @@
 use crate::{scanner::token::Token, utils::error::Error};
 
-use super::expr::{self, Expr};
+use super::Expr;
 
 pub struct Binary<R> {
     left: Box<dyn Expr<R>>,
@@ -31,7 +31,7 @@ impl<R> Binary<R> {
 }
 
 impl<R: 'static> Expr<R> for Binary<R> {
-    fn accept(&self, visitor: &mut dyn expr::Visitor<R>) -> Result<R, Error> {
+    fn accept(&self, visitor: &mut dyn super::Visitor<R>) -> Result<R, Error> {
         visitor.visit_binary_expr(self)
     }
