+++ interpreter/src/ast/expr/grouping.rs
@@ -1,6 +1,6 @@
 use crate::utils::error::Error;
 
-use super::expr::{self, Expr};
+use super::Expr;
 
 pub struct Grouping<R> {
     expression: Box<dyn Expr<R>>,
@@ -17,7 +17,7 @@ impl<R> Grouping<R> {
 }
 
 impl<R: 'static> Expr<R> for Grouping<R> {
-    fn accept(&self, visitor: &mut dyn expr::Visitor<R>) -> Result<R, Error> {
+    fn accept(&self, visitor: &mut dyn super::Visitor<R>) -> Result<R, Error> {
         visitor.visit_grouping_expr(self)
     }
