+++ interpreter/src/ast/expr/variable.rs
@@ -1,6 +1,6 @@
 use crate::{scanner::token::Token, utils::error::Error};
 
-use super::expr::{self, Expr};
+use super::Expr;
 
 pub struct Variable {
     name: Token,
@@ -17,7 +17,7 @@ impl Variable {
 }
 
 impl<R: 'static> Expr<R> for Variable {
-    fn accept(&self, visitor: &mut dyn expr::Visitor<R>) -> Result<R, Error> {
+    fn accept(&self, visitor: &mut dyn super::Visitor<R>) -> Result<R, Error> {
         visitor.visit_variable_expr(self)
     }
