+++ interpreter/src/parser/mod.rs
@@ -1,12 +1,7 @@
 mod error;
 
 use crate::{
-    ast::{
-        assign::Assign, binary::Binary, block::Block, expr::Expr, expression::Expression,
-        grouping::Grouping, literal::Literal, logical::Logical, print::Print, r#if::If,
-        r#while::While, stmt::Stmt, unary::Unary, var::Var, variable::Variable,
-    },
-    scanner::{
+    ast::{expr::{assign::Assign, binary::Binary, call::Call, grouping::Grouping, literal::Literal, logical::Logical, unary::Unary, variable::Variable, Expr}, stmt::{block::Block, expression::Expression, function::Function, r#if::If, print::Print, var::Var, r#while::While, Stmt}}, parser::error::error, scanner::{
         token::{Object, Token},
         token_type::TokenType,
     },
